# docs/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "StayBuddy API",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/v1"
    }
  ],
  "paths": {
...
    "/mobile/bootstrap": {
      "get": {
        "operationId": "getMobileBootstrap",
        "parameters": [
          {
            "name": "X-App-Installation-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SignedBootstrapManifest",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignedBootstrapManifest"
                }
              }
            }
          },
          "404": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/ops/hotels": {
      "get": {
        "operationId": "listHotels",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "HotelSummary",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HotelSummary"
                  }
                }
              }
            }
          },
          "401": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createHotel",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHotelInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "CreatedHotel",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedHotel"
                }
              }
            }
          },
          "400": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/ops/app-builds": {
      "post": {
        "operationId": "createAppBuild",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppBuildCommand"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "AppBuildReceipt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppBuildReceipt"
                }
              }
            }
          },
          "400": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/reservation-imports/preview": {
      "post": {
        "operationId": "previewReservationImport",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservationPreviewInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ImportPreview",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportPreview"
                }
              }
            }
          },
          "400": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/reservation-imports/commit": {
      "post": {
        "operationId": "commitReservationImport",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservationCommitInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "ReservationCommitResult",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationCommitResult"
                }
              }
            }
          },
          "400": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/reservations/manual": {
      "post": {
        "operationId": "createManualReservation",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CanonicalReservation"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "ReservationCommitResult",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationCommitResult"
                }
              }
            }
          },
          "400": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/reservations": {
      "get": {
        "operationId": "listReservations",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "ReservationSummary",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReservationSummary"
                  }
                }
              }
            }
          },
          "401": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "INVALID_REQUEST",
              "UNAUTHENTICATED",
              "FORBIDDEN",
              "TENANT_NOT_RESOLVED",
              "NOT_FOUND",
              "CONFLICT",
              "RATE_LIMITED",
              "CLAIM_EXPIRED",
              "CLAIM_REPLAYED",
              "CLAIM_REVOKED",
              "OTP_INVALID",
              "OTP_EXPIRED",
              "TERMS_REQUIRED",
              "INTERNAL_ERROR"
            ]
          },
          "traceId": {
            "type": "string",
            "minLength": 8,
            "maxLength": 80
          },
          "retryable": {
            "type": "boolean"
          },
          "metadata": {
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {
              "type": [
                "string",
                "number",
                "boolean"
              ]
            }
          }
        },
        "required": [
          "code",
          "traceId",
          "retryable"
        ],
        "additionalProperties": false
      },
...
      "SignedBootstrapManifest": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "manifest": {
            "type": "object",
            "properties": {
              "schemaVersion": {
                "type": "number",
                "const": 1
              },
              "appInstallationKey": {
                "type": "string",
                "minLength": 16,
                "maxLength": 120
              },
              "hotelId": {
                "type": "string",
                "minLength": 8,
                "maxLength": 80
              },
              "appId": {
                "type": "string",
                "minLength": 8,
                "maxLength": 80
              },
              "appName": {
                "type": "string",
                "minLength": 1,
                "maxLength": 80
              },
              "hotelDisplayName": {
                "type": "string",
                "minLength": 1,
                "maxLength": 120
              },
              "theme": {
                "type": "object",
                "properties": {
                  "primary": {
                    "type": "string",
                    "pattern": "^#[0-9A-F]{6}$"
                  },
                  "accent": {
                    "type": "string",
                    "pattern": "^#[0-9A-F]{6}$"
                  },
                  "canvas": {
                    "type": "string",
                    "pattern": "^#[0-9A-F]{6}$"
                  },
                  "surfaceWarm": {
                    "type": "string",
                    "pattern": "^#[0-9A-F]{6}$"
                  },
                  "ink": {
                    "type": "string",
                    "pattern": "^#[0-9A-F]{6}$"
                  },
                  "divider": {
                    "type": "string",
                    "pattern": "^#[0-9A-F]{6}$"
                  },
                  "logoUrl": {
                    "type": "string",
                    "format": "uri"
                  },
                  "heroImageUrl": {
                    "type": "string",
                    "format": "uri"
                  }
                },
                "required": [
                  "primary",
                  "accent",
                  "canvas",
                  "surfaceWarm",
                  "ink",
                  "divider",
                  "logoUrl"
                ],
                "additionalProperties": false
              },
              "supportedLocales": {
                "minItems": 4,
                "maxItems": 4,
                "type": "array",
                "items": {
                  "type": "string",
                  "enum": [
                    "en",
                    "th",
                    "zh-CN",
                    "ru"
                  ]
                }
              },
              "defaultLocale": {
                "type": "string",
                "enum": [
                  "en",
                  "th",
                  "zh-CN",
                  "ru"
                ]
              },
              "voiceProfile": {
                "type": "string",
                "enum": [
                  "FIVE_STAR_RESORT",
                  "FIVE_STAR_BOUTIQUE"
                ]
              },
              "features": {
                "type": "object",
                "propertyNames": {
                  "type": "string"
                },
                "additionalProperties": {
                  "type": "boolean"
                }
              },
              "minimumVersion": {
                "type": "string",
                "pattern": "^\\d+\\.\\d+\\.\\d+$"
              },
              "maintenance": {
                "type": "object",
                "properties": {
                  "active": {
                    "type": "boolean"
                  },
                  "messageKey": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "active"
                ],
                "additionalProperties": false
              },
              "issuedAt": {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
              },
              "expiresAt": {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
              }
            },
            "required": [
              "schemaVersion",
              "appInstallationKey",
              "hotelId",
              "appId",
              "appName",
              "hotelDisplayName",
              "theme",
              "supportedLocales",
              "defaultLocale",
              "voiceProfile",
              "features",
              "minimumVersion",
              "maintenance",
              "issuedAt",
              "expiresAt"
            ],
            "additionalProperties": false
          },
          "signature": {
            "type": "string",
            "pattern": "^[0-9a-f]{128}$"
          },
          "algorithm": {
            "type": "string",
            "const": "Ed25519"
          }
        },
        "required": [
          "manifest",
          "signature",
          "algorithm"
        ],
        "additionalProperties": false
      },
...
      "CreateHotelInput": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "legalName": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "roomCount": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991
          },
          "timezone": {
            "type": "string"
          },
          "countryCode": {
            "type": "string",
            "minLength": 2,
            "maxLength": 2
          }
        },
        "required": [
          "slug",
          "legalName",
          "displayName",
          "roomCount",
          "timezone"
        ],
        "additionalProperties": false
      },
      "CreatedHotel": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "hotelId": {
            "type": "string",
            "minLength": 8,
            "maxLength": 80
          },
          "appId": {
            "type": "string",
            "minLength": 8,
            "maxLength": 80
          },
          "appInstallationKey": {
            "type": "string",
            "minLength": 16
          },
          "status": {
            "type": "string",
            "const": "ONBOARDING"
          },
          "nextStep": {
            "type": "string"
          }
        },
        "required": [
          "hotelId",
          "appId",
          "appInstallationKey",
          "status",
          "nextStep"
        ],
        "additionalProperties": false
      },
      "HotelSummary": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 8,
            "maxLength": 80
          },
          "slug": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "roomCount": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991
          },
          "appStatus": {
            "type": [
              "string",
              "null"
            ]
          },
          "appUpdatedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "slug",
          "displayName",
          "status",
          "roomCount",
          "appStatus",
          "appUpdatedAt"
        ],
        "additionalProperties": false
      },
      "AppBuildCommand": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "hotelId": {
            "type": "string",
            "minLength": 8,
            "maxLength": 80
          },
          "hotelAppId": {
            "type": "string",
            "minLength": 8,
            "maxLength": 80
          },
          "platform": {
            "type": "string",
            "enum": [
              "IOS",
              "ANDROID"
            ]
          },
          "profile": {
            "type": "string",
            "enum": [
              "DEVELOPMENT",
              "PREVIEW",
              "PRODUCTION"
            ]
          },
          "version": {
            "type": "string"
          }
        },
        "required": [
          "hotelId",
          "hotelAppId",
          "platform",
          "profile",
          "version"
        ],
        "additionalProperties": false
      },
      "AppBuildReceipt": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "buildJobId": {
            "type": "string",
            "minLength": 8,
            "maxLength": 80
          },
          "status": {
            "type": "string",
            "const": "QUEUED"
          }
        },
        "required": [
          "buildJobId",
          "status"
        ],
        "additionalProperties": false
      },
...
      "ReservationPreviewInput": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "csv": {
            "type": "string",
            "minLength": 1
          },
          "mapping": {
            "type": "object",
            "properties": {
              "sourceSystem": {
                "type": "string",
                "minLength": 1,
                "maxLength": 80
              },
              "columns": {
                "type": "object",
                "properties": {
                  "externalReservationId": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string"
                  },
                  "bookingSource": {
                    "type": "string"
                  },
                  "confirmationCode": {
                    "type": "string"
                  },
                  "guestName": {
                    "type": "string"
                  },
                  "guestEmail": {
                    "type": "string"
                  },
                  "nationality": {
                    "type": "string"
                  },
                  "preferredLanguage": {
                    "type": "string"
                  },
                  "checkInAt": {
                    "type": "string"
                  },
                  "checkOutAt": {
                    "type": "string"
                  },
                  "roomType": {
                    "type": "string"
                  },
                  "roomNumber": {
                    "type": "string"
                  },
                  "adults": {
                    "type": "string"
                  },
                  "children": {
                    "type": "string"
                  },
                  "updatedAtSource": {
                    "type": "string"
                  }
                },
                "required": [
                  "externalReservationId",
                  "status",
                  "bookingSource",
                  "confirmationCode",
                  "guestName",
                  "checkInAt",
                  "checkOutAt"
                ],
                "additionalProperties": false
              },
              "defaults": {
                "type": "object",
                "properties": {
                  "timezone": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 80
                  },
                  "bookingSource": {
                    "type": "string"
                  }
                },
                "required": [
                  "timezone"
                ],
                "additionalProperties": false
              }
            },
            "required": [
              "sourceSystem",
              "columns",
              "defaults"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "csv",
          "mapping"
        ],
        "additionalProperties": false
      },
...
      "ReservationCommitResult": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "batchId": {
            "type": "string",
            "minLength": 8,
            "maxLength": 80
          },
          "status": {
            "type": "string",
            "enum": [
              "COMPLETED",
              "PARTIALLY_REJECTED"
            ]
          },
          "created": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "updated": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "rejected": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "batchId",
          "status",
          "created",
          "updated",
          "rejected"
        ],
        "additionalProperties": false
      },
      "ReservationSummary": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 8,
            "maxLength": 80
          },
          "externalReservationId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "bookingSource": {
            "type": "string"
          },
          "confirmationCode": {
            "type": "string"
          },
          "primaryGuestName": {
            "type": "string"
          },
          "nationality": {
            "type": [
              "string",
              "null"
            ]
          },
          "preferredLocale": {
            "type": [
              "string",
              "null"
            ]
          },
          "checkInAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          },
          "checkOutAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          },
          "stayId": {
            "type": "string",
            "minLength": 8,
            "maxLength": 80
          },
          "lifecycle": {
            "type": "string"
          },
          "roomType": {
            "type": [
              "string",
              "null"
            ]
          },
          "roomNumber": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "externalReservationId",
          "status",
          "bookingSource",
          "confirmationCode",
          "primaryGuestName",
          "nationality",
          "preferredLocale",
          "checkInAt",
          "checkOutAt",
          "stayId",
          "lifecycle",
          "roomType",
          "roomNumber"
        ],
        "additionalProperties": false
      },
...
    },
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 8,
          "maxLength": 200
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}